Split into separate "filer" crate.
