+++ Cargo.toml
@@ -3,7 +3,7 @@
 debug = true
 
 [profile.release]
-lto = true
+#lto = true
 debug = true
 
 [workspace]
@@ -14,6 +14,7 @@ members = [
     "memflow-framework",
     "memflow-daemon-connector",
     "memflow-node",
+	"filer",
 ]
 default-members = [
     "memflow-daemon",
@@ -22,4 +23,5 @@ default-members = [
     "memflow-framework",
     "memflow-daemon-connector",
     "memflow-node",
+	"filer",
 ]
