+++ filer/Cargo.toml
@@ -1,12 +1,17 @@
 [package]
-name = "memflow-framework"
+name = "filer"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-memflow = { git = "https://github.com/memflow/memflow", branch = "next" }
-abi_stable = "0.10"
+abi_stable = { version = "0.10" }
 cglue = { version = ">=0.2.5", git = "https://github.com/h33p/cglue", branch = "main", default-features = false, features = ["layout_checks"] }
-filer = { version = "0.1", path = "../filer" }
+dashmap = "5"
+sharded-slab = "0.1"
+crossbeam-deque = "0.8"
+
+[features]
+default = ["std"]
+std = []
