+++ filer/src/lib.rs
@@ -1,0 +1,17 @@
+pub mod backend;
+pub mod branch;
+pub mod error;
+pub mod fs;
+pub mod node;
+pub mod plugin_store;
+pub mod thread_ctx;
+pub mod types;
+
+pub mod prelude {
+    pub mod v1 {
+        pub use crate::{
+            backend::*, branch::*, error::*, fs::*, node::*, plugin_store::*, thread_ctx::*,
+            types::*,
+        };
+    }
+}
