+++ memflow-node/Cargo.toml
@@ -12,3 +12,4 @@ ptree = "0.4"
 rand = "0.8"
 rand_xorshift = "0.3"
 memflow = { git = "https://github.com/memflow/memflow", branch = "next" }
+filer = { version = "0.1", path = "../filer" }
